MDP-203: rename footer component
