+++ src/components/Footer/Info.tsx
